requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.3
groq==0.4.2
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1 